requirements.txt
annotated-types==0.7.0
anyio==4.5.2
bip39==0.0.2
attrs==25.3.0
base58==2.1.1
borsh-construct==0.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
construct==2.10.68
construct-typing==0.5.6
ed25519==1.5
exceptiongroup==1.2.2
fastapi>=0.115.12
h11==0.14.0
idna==3.10
near-api-py==0.1.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
starlette==0.44.0
sumtypes==0.1a6
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn>=0.33.0
